% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_device}
\alias{get_device}
\title{The only distinction is between mobile and desktop. If the user is on a tablet,
the results are ambiguous; some tablets report as "Mobile" while some tablets
report as "Desktop". look at width}
\usage{
get_device()
}
\description{
The only distinction is between mobile and desktop. If the user is on a tablet,
the results are ambiguous; some tablets report as "Mobile" while some tablets
report as "Desktop". look at width
}
