% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{is_browser_firefox}
\alias{is_browser_firefox}
\title{Is the user using Firefox?}
\usage{
is_browser_firefox()
}
\value{
Whether or not this user using Firefox
}
\description{
Convenience function that checks if the user's browser is detected as Firefox.
See \code{\link[=get_browser]{get_browser()}} for details.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    shinybrowser::detect(),
    "Are you using Firefox?",
    textOutput("result")
  )
  server <- function(input, output, session) {
    output$result <- renderText({
      shinybrowser::is_browser_firefox()
    })
  }
  shinyApp(ui, server)
}
}
