% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{is_device_mobile}
\alias{is_device_mobile}
\title{Is the user on a mobile device?}
\usage{
is_device_mobile()
}
\description{
Convenience function that checks if the user's device is detected as mobile.
See \code{\link[=get_device]{get_device()}} for details.
}
