% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{is_os_mac}
\alias{is_os_mac}
\title{Is the user on Mac?}
\usage{
is_os_mac()
}
\description{
Convenience function that checks if the user's operating system is detected as Mac.
See \code{\link[=get_os]{get_os()}} for details.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    shinybrowser::detect(),
    "Are you on Mac?",
    textOutput("result")
  )
  server <- function(input, output, session) {
    output$result <- renderText({
      shinybrowser::is_os_mac()
    })
  }
  shinyApp(ui, server)
}
}
