% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybrowser-package.R
\docType{package}
\name{shinybrowser-package}
\alias{shinybrowser}
\alias{shinybrowser-package}
\title{shinybrowser: Find Out Information About a User's Web Browser in 'Shiny'}
\description{
Sometimes it's useful to know some information about your user in a 'Shiny' app. The available information is: browser name (such as 'Chrome' or 'Safari') and version, device type (mobile or desktop), operating system (such as 'Windows' or 'Mac' or 'Android') and version, and browser dimensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/shinybrowser}
  \item \url{https://daattali.com/shiny/shinybrowser-demo/}
  \item Report bugs at \url{https://github.com/daattali/shinybrowser/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

}
\keyword{internal}
