% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\docType{data}
\name{SUPPORTED_BROWSERS}
\alias{SUPPORTED_BROWSERS}
\title{Browsers that can be detected with \{shinybrowser\}}
\description{
Browsers that can be detected with \{shinybrowser\}
}
\keyword{datasets}
