% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\docType{data}
\name{SUPPORTED_DEVICES}
\alias{SUPPORTED_DEVICES}
\title{Devices that can be detected with \{shinybrowser\}}
\description{
Devices that can be detected with \{shinybrowser\}
}
\keyword{datasets}
