% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\docType{data}
\name{SUPPORTED_OPERATING_SYSTEMS}
\alias{SUPPORTED_OPERATING_SYSTEMS}
\title{Operating systems that can be detected with \{shinybrowser\}}
\description{
Operating systems that can be detected with \{shinybrowser\}
}
\keyword{datasets}
