% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{Detect a user's browser information}
\usage{
detect()
}
\description{
This function must be called somewhere in a Shiny app's UI in order to use any other
\{shinybrowser\} functions.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    shinybrowser::detect(),
    "Your browser information:",
    verbatimTextOutput("browser_info")
  )
  server <- function(input, output, session) {
    output$browser_info <- renderPrint({
      shinybrowser::get_info()
    })
  }
  shinyApp(ui, server)
}
}
\seealso{
\code{\link[=get_info]{get_info()}}
}
