% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{is_browser_ie}
\alias{is_browser_ie}
\title{Is the user using Internet Explorer?}
\usage{
is_browser_ie()
}
\description{
Convenience function that checks if the user's browser is detected as Internet Explorer.
See \code{\link[=get_browser]{get_browser()}} for details.
}
